+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 
+## v0.2.13
+
+* handle large emit_configuration to database emitter by checking the size of the packet, and removing process parameters if necessary.
+
+## v0.2.12
+
+* plotting tweaks. 
+* provide information about update failure.
+* improve data_from_database to return experiment config.
+
 ## v0.2.11
 
 * process.schema = None in base class constructor